fix: 修复查询用户自己的作业

+++ src/main/java/plus/maa/backend/controller/CopilotController.java
@@ -1,6 +1,7 @@
 package plus.maa.backend.controller;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.util.ObjectUtils;
 import org.springframework.web.bind.annotation.*;
 import plus.maa.backend.common.annotation.CurrentUser;
 import plus.maa.backend.controller.request.CopilotCUDRequest;
@@ -8,6 +9,7 @@ import plus.maa.backend.controller.request.CopilotQueriesRequest;
 import plus.maa.backend.controller.response.CopilotInfo;
 import plus.maa.backend.controller.response.CopilotPageInfo;
 import plus.maa.backend.controller.response.MaaResult;
+import plus.maa.backend.controller.response.MaaResultException;
 import plus.maa.backend.service.CopilotService;
 import plus.maa.backend.service.model.LoginUser;
 
@@ -25,12 +27,14 @@ public class CopilotController {
     @PostMapping("/upload")
     public MaaResult<String> uploadCopilot(@CurrentUser LoginUser loginUser,
                                            @RequestBody CopilotCUDRequest request) {
+        if (ObjectUtils.isEmpty(loginUser)) throw new MaaResultException("请先登录账号");
         return copilotService.upload(loginUser, request.getContent());
     }
 
     @PostMapping("/delete")
     public MaaResult<Void> deleteCopilot(@CurrentUser LoginUser loginUser,
                                          @RequestBody CopilotCUDRequest request) {
+        if (ObjectUtils.isEmpty(loginUser)) throw new MaaResultException("请先登录账号");
         return copilotService.delete(loginUser, request);
     }
 
@@ -41,18 +45,20 @@ public class CopilotController {
 
 
     @GetMapping("/query")
-    public MaaResult<CopilotPageInfo> queriesCopilot(CopilotQueriesRequest request) {
-        return copilotService.queriesCopilot(request);
+    public MaaResult<CopilotPageInfo> queriesCopilot(@CurrentUser LoginUser loginUser, CopilotQueriesRequest request) {
+        return copilotService.queriesCopilot(loginUser, request);
     }
 
     @PostMapping("/update")
     public MaaResult<Void> updateCopilot(@CurrentUser LoginUser loginUser,
                                          @RequestBody CopilotCUDRequest request) {
+        if (ObjectUtils.isEmpty(loginUser)) throw new MaaResultException("请先登录账号");
         return copilotService.update(loginUser, request.getId(), request.getContent());
     }
 
     @PostMapping("/rating")
-    public MaaResult<Void> ratesCopilotOperation(CopilotQueriesRequest request) {
+    public MaaResult<Void> ratesCopilotOperation(@CurrentUser LoginUser loginUser,CopilotQueriesRequest request) {
+        if (ObjectUtils.isEmpty(loginUser)) throw new MaaResultException("请先登录账号");
         return null;
     }
